move thresholds array to generator.options
