+++ src/codegen/codegen.test.ts
@@ -84,6 +84,7 @@ describe('Code generation', () => {
                   value: 1,
                 },
               },
+              thresholds: [],
             },
             testData: {
               variables: [],
@@ -93,7 +94,6 @@ describe('Code generation', () => {
             allowlist: [],
             includeStaticAssets: false,
             scriptName: 'my-script.js',
-            thresholds: [],
           },
         }).replace(/\s/g, '')
       ).toBe(expectedResult.replace(/\s/g, ''))
