+++ src/schemas/generator/v1/index.ts
@@ -2,7 +2,6 @@ import { z } from 'zod'
 import { TestRuleSchema } from './rules'
 import { TestDataSchema } from './testData'
 import { TestOptionsSchema } from './testOptions'
-import { ThresholdSchema } from './thresholds'
 
 export const GeneratorFileDataSchema = z.object({
   version: z.literal('1.0'),
@@ -13,7 +12,6 @@ export const GeneratorFileDataSchema = z.object({
   allowlist: z.string().array(),
   includeStaticAssets: z.boolean(),
   scriptName: z.string().default('my-script.js'),
-  thresholds: z.array(ThresholdSchema).default([]),
 })
 
 export type GeneratorSchema = z.infer<typeof GeneratorFileDataSchema>
