+++ src/store/generator/useGeneratorStore.ts
@@ -15,15 +15,13 @@ import { exhaustive } from '@/utils/typescript'
 import { GeneratorFileData } from '@/types/generator'
 import { ProxyData } from '@/types'
 import { createScriptDataSlice, ScriptDataStore } from './slices/script'
-import { createThresholdSlice, ThresholdSliceStore } from './slices/thresholds'
 
 export interface GeneratorStore
   extends RecordingSliceStore,
     RulesSliceStore,
     TestDataStore,
     TestOptionsStore,
-    ScriptDataStore,
-    ThresholdSliceStore {
+    ScriptDataStore {
   setGeneratorFile: (
     generatorFile: GeneratorFileData,
     recording?: ProxyData[]
@@ -37,17 +35,15 @@ export const useGeneratorStore = create<GeneratorStore>()(
     ...createTestDataSlice(set, ...rest),
     ...createTestOptionsSlice(set, ...rest),
     ...createScriptDataSlice(set, ...rest),
-    ...createThresholdSlice(set, ...rest),
     setGeneratorFile: (
       {
-        options: { thinkTime, loadProfile },
+        options: { thinkTime, loadProfile, thresholds },
         testData: { variables, files },
         recordingPath,
         rules,
         allowlist,
         includeStaticAssets,
         scriptName,
-        thresholds,
       },
       recording = []
     ) =>
@@ -55,6 +51,7 @@ export const useGeneratorStore = create<GeneratorStore>()(
         // options
         state.sleepType = thinkTime.sleepType
         state.timing = thinkTime.timing
+        state.thresholds = thresholds
         state.executor = loadProfile.executor
         switch (loadProfile.executor) {
           case 'ramping-vus':
@@ -83,8 +80,6 @@ export const useGeneratorStore = create<GeneratorStore>()(
         state.scriptName = scriptName
         // rules
         state.rules = rules
-        // thresholds
-        state.thresholds = thresholds
       }),
   }))
 )
