+++ src/types/testOptions.ts
@@ -11,6 +11,10 @@ import {
   RampingStageSchema,
   LoadProfileExecutorOptionsSchema,
   TestOptionsSchema,
+  ThresholdDataSchema,
+  ThresholdMetricSchema,
+  ThresholdSchema,
+  ThresholdStatisticSchema,
 } from '@/schemas/generator'
 
 export type SleepType = z.infer<typeof SleepTypeSchema>
@@ -27,4 +31,8 @@ export type RampingVUsOptions = z.infer<typeof RampingVUsOptionsSchema>
 export type LoadProfileExecutorOptions = z.infer<
   typeof LoadProfileExecutorOptionsSchema
 >
+export type Threshold = z.infer<typeof ThresholdSchema>
+export type ThresholdData = z.infer<typeof ThresholdDataSchema>
+export type ThresholdMetric = z.infer<typeof ThresholdMetricSchema>
+export type ThresholdStatstic = z.infer<typeof ThresholdStatisticSchema>
 export type TestOptions = z.infer<typeof TestOptionsSchema>
