+++ src/views/Generator/TestOptions/Thresholds/ThresholdRow.tsx
@@ -1,5 +1,4 @@
 import { FieldGroup, ControlledSelect } from '@/components/Form'
-import { Threshold, ThresholdData } from '@/types/thresholds'
 import { TrashIcon } from '@radix-ui/react-icons'
 import { Table, TextField, Checkbox, IconButton, Flex } from '@radix-ui/themes'
 import {
@@ -18,6 +17,7 @@ import { useThresholdURLOptions } from './Thresholds.hooks'
 import { css } from '@emotion/react'
 import { useTheme } from '@/hooks/useTheme'
 import { useEffect } from 'react'
+import { ThresholdData, Threshold } from '@/types/testOptions'
 
 type ThresholdRowProps = {
   index: number
