+++ src/views/Generator/TestOptions/Thresholds/Thresholds.tsx
@@ -1,6 +1,5 @@
 import { ThresholdDataSchema } from '@/schemas/generator'
 import { useGeneratorStore } from '@/store/generator'
-import { Threshold, ThresholdData } from '@/types/thresholds'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 import { Button, Text } from '@radix-ui/themes'
@@ -8,6 +7,7 @@ import { useCallback, useEffect } from 'react'
 import { useForm, useFieldArray, FormProvider } from 'react-hook-form'
 import { ThresholdRow } from './ThresholdRow'
 import { Table } from '@/components/Table'
+import { Threshold, ThresholdData } from '@/types/testOptions'
 
 export function Thresholds() {
   const thresholds = useGeneratorStore((store) => store.thresholds)
